+++ spec/q-spec.js
@@ -196,9 +196,9 @@ describe("always next tick", function () {
 		spyOn(Q, 'nextTick').andCallFake(function immediateTick(task){
 			task();
 		});
-		
+
 		Q.when(Q(), spy);
-		
+
 		expect(spy).toHaveBeenCalled();
 		expect(Q.nextTick).toHaveBeenCalled();
 	});
@@ -1183,6 +1183,223 @@ describe("all", function () {
 
 });
 
+describe("any", function() {
+    it("fulfills when passed an empty array", function() {
+        return Q.any([]);
+    });
+
+    it("rejects after all promises are rejected", function() {
+        var deferreds = [Q.defer(), Q.defer()];
+        var promises = [deferreds[0].promise, deferreds[1].promise];
+
+        return testReject(promises, deferreds);
+    });
+
+    it("rejects after all promises in a sparse array are rejected", function() {
+        var deferreds = [Q.defer(), Q.defer()];
+        var promises = [];
+        promises[0] = deferreds[0].promise;
+        promises[3] = deferreds[1].promise;
+
+        return testReject(promises, deferreds);
+    });
+
+    function testReject(promises, deferreds) {
+        var promise = Q.any(promises);
+
+        for (var index = 0; index < deferreds.length; index++) {
+            var deferred = deferreds[index];
+            (function() {
+                deferred.reject(new Error('Rejected'));
+            })();
+        }
+
+        return Q.delay(250)
+          .then(function() {
+              expect(promise.isRejected()).toBe(true);
+              expect(promise.inspect().reason.message)
+              .toBe("Can't get fulfillment value from any promise, all promises were rejected.");
+          })
+          .timeout(1000);
+    }
+
+    it("fulfills with the first resolved promise", function() {
+        var deferreds = [Q.defer(), Q.defer()];
+        var promises = [deferreds[0].promise, deferreds[1].promise];
+
+        testFulfill(promises, deferreds);
+    });
+
+    it("fulfills when passed a sparse array", function() {
+        var deferreds = [Q.defer(), Q.defer()];
+        var promises = [];
+        promises[0] = deferreds[0].promise;
+        promises[2] = deferreds[1].promise;
+
+        testFulfill(promises, deferreds);
+    });
+
+    function testFulfill(promises, deferreds) {
+        var promise = Q.any(promises);
+
+        var j = 1;
+        for (var index = 0; index < deferreds.length; index++) {
+            var toResolve = deferreds[index];
+            if (!toResolve || !Q.isPromiseAlike(toResolve.promise)) {
+                continue;
+            }
+
+            (function(index, toResolve) {
+                var time = index * 50;
+                Q.delay(time).then(function() {
+                    toResolve.resolve('Fulfilled' + index);
+                });
+            })(j, toResolve);
+
+            j++;
+        }
+
+        return Q.delay(400)
+          .then(function() {
+              expect(promise.isFulfilled()).toBe(true);
+              expect(promise.inspect().value).toBe('Fulfilled1');
+          })
+          .timeout(1000);
+    }
+
+    it("fulfills with the first value", function() {
+        var toResolve1 = Q.defer();
+        var toResolve2 = Q.defer();
+        var toResolve3 = Q.defer();
+        var promises = [toResolve1.promise, toResolve2.promise, 4, 5,
+            toResolve3.promise
+        ];
+
+        var promise = Q.any(promises);
+
+        Q.delay(150).then(function() {
+            toResolve1.resolve(1);
+        });
+        Q.delay(50).then(function() {
+            toResolve2.resolve(2);
+        });
+        Q.delay(100).then(function() {
+            toResolve3.resolve(3);
+        });
+
+        return Q.delay(250)
+          .then(function() {
+              expect(promise.isFulfilled()).toBe(true);
+              expect(promise.inspect().value).toBe(4);
+          })
+          .timeout(1000);
+    });
+
+    it("fulfills after rejections", function() {
+        var toReject = [Q.defer(), Q.defer()];
+        var toResolve = Q.defer();
+        var promises = [toReject[0].promise, toReject[1].promise,
+            toResolve
+            .promise
+        ];
+
+        var promise = Q.any(promises);
+
+        testFulfillAfterRejections(promises, toReject, toResolve);
+    });
+
+    it("fulfills after rejections in sparse array", function() {
+        var toReject = [Q.defer(), Q.defer()];
+        var toResolve = Q.defer();
+        var promises = [];
+        promises[2] = toReject[0].promise;
+        promises[5] = toReject[1].promise;
+        promises[9] = toResolve.promise;
+
+        testFulfillAfterRejections(promises, toReject, toResolve);
+    });
+
+    function testFulfillAfterRejections(promises, rejectDeferreds,
+      fulfillDeferred) {
+        var promise = Q.any(promises);
+
+        for (var index = 0; index < rejectDeferreds.length; index++) {
+            var toReject = rejectDeferreds[index];
+            (function(index, toReject) {
+                var time = (index + 1) * 50;
+                Q.delay(time).then(function() {
+                    toReject.reject(new Error('Rejected'));
+                });
+            })(index, toReject);
+
+            index++;
+        }
+        Q.delay(index * 50).then(function() {
+            fulfillDeferred.resolve('Fulfilled');
+        });
+
+        return Q.delay(400)
+          .then(function() {
+              expect(promise.isFulfilled()).toBe(true);
+              expect(promise.inspect().value).toBe('Fulfilled');
+          })
+          .timeout(1000);
+    }
+
+    it("resolves foreign thenables", function() {
+        var normal = Q.delay(150)
+            .then(function() {})
+            .thenResolve(1);
+        var foreign = {
+            then: function(f) {
+                return f(2);
+            }
+        };
+
+        return Q.any([normal, foreign])
+          .then(function(result) {
+              expect(result).toEqual(2);
+          });
+    });
+
+    it("sends { index, value } progress updates", function() {
+        var deferred1 = Q.defer();
+        var deferred2 = Q.defer();
+
+        var progressValues = [];
+
+        Q.delay(50).then(function() {
+            deferred1.notify("a");
+        });
+        Q.delay(100).then(function() {
+            deferred2.notify("b");
+            deferred2.resolve();
+        });
+        Q.delay(150).then(function() {
+            deferred1.notify("c"); // Is lost, deferred2 already resolved.
+            deferred1.resolve();
+        });
+
+        return Q.any([deferred1.promise, deferred2.promise])
+          .delay(250)
+          .then(function() {
+              expect(progressValues).toEqual([{
+                  index: 0,
+                  value: "a"
+              }, {
+                  index: 1,
+                  value: "b"
+              }]);
+            },
+            undefined,
+            function(progressValue) {
+                progressValues.push(progressValue);
+            }
+          );
+    });
+
+});
+
 describe("allSettled", function () {
     it("works on an empty array", function () {
         return Q.allSettled([])
@@ -2117,7 +2334,7 @@ describe("node support", function () {
         });
 
     });
-    
+
 	describe("npost", function () {
 
         it("fulfills with callback result", function () {
@@ -2269,7 +2486,7 @@ describe("node support", function () {
                 expect(ten).toBe(10);
             });
         });
-		
+
     });
 
 });